requirements.txt
telethon
sqlalchemy
black
pylint
python-dotenv
pytest-asyncio
pytest-env
pytest-mock
asyncclick
anyio
